+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/Quotation/QuotationDTOConverter.java
@@ -1,4 +1,4 @@
-package com.kakao.sunsuwedding.match.Quotation;
+package com.kakao.sunsuwedding.Quotation;
 
 import java.util.List;
