+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/Quotation/QuotationRequest.java
@@ -1,4 +1,4 @@
-package com.kakao.sunsuwedding.match.Quotation;
+package com.kakao.sunsuwedding.Quotation;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
