+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/Quotation/QuotationResponse.java
@@ -1,4 +1,4 @@
-package com.kakao.sunsuwedding.match.Quotation;
+package com.kakao.sunsuwedding.Quotation;
 
 import java.time.LocalDateTime;
 import java.util.List;
