+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/match/MatchService.java
@@ -5,9 +5,9 @@ import com.kakao.sunsuwedding._core.errors.exception.BadRequestException;
 import com.kakao.sunsuwedding._core.errors.exception.ForbiddenException;
 import com.kakao.sunsuwedding._core.errors.exception.NotFoundException;
 import com.kakao.sunsuwedding._core.utils.PriceCalculator;
-import com.kakao.sunsuwedding.match.Quotation.Quotation;
-import com.kakao.sunsuwedding.match.Quotation.QuotationJPARepository;
-import com.kakao.sunsuwedding.match.Quotation.QuotationStatus;
+import com.kakao.sunsuwedding.Quotation.Quotation;
+import com.kakao.sunsuwedding.Quotation.QuotationJPARepository;
+import com.kakao.sunsuwedding.Quotation.QuotationStatus;
 import com.kakao.sunsuwedding.portfolio.PortfolioService;
 import com.kakao.sunsuwedding.user.constant.Role;
 import com.kakao.sunsuwedding.user.couple.Couple;
